requirements.txt
flask>=3.0,<4
requests>=2.31
beautifulsoup4>=4.12
lxml>=5.0
pymupdf>=1.24
pytesseract>=0.3.10
Pillow>=10.0
ezdxf>=1.3
pywebview>=5.0
pystray>=0.19
bcrypt>=4.0
itsdangerous>=2.2
stripe>=8.0
pytest>=8.0
